% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_module.R
\name{get_pivot_vars}
\alias{get_pivot_vars}
\title{Create a table with one row per pivot variable}
\usage{
get_pivot_vars(df, max_levels = 1000)
}
\arguments{
\item{df}{A local dataframe or tbl_dbi database table}

\item{max_levels}{The maximum number of levels a pivot variable is allowed to have.}
}
\value{
A tibble with one row per pivot variable with the variable's name, number of levels, and a list column containing the levels.
}
\description{
Create a table with one row per pivot variable
}
