% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_module.R
\name{pivot_module_UI}
\alias{pivot_module_UI}
\title{Generate the UI for a pivot table}
\usage{
pivot_module_UI(id, pivot_vars, sum_vars)
}
\arguments{
\item{id}{The namespace id as a string. Can be anything but must match the corresponding namespace id of pivot_module.}

\item{pivot_vars}{A tibble created by get_pivot_vars()}

\item{sum_vars}{A character vector contiaining the names of numeric variables in the data to summarise by.}
}
\value{
A tag list containing the UI elements for a pivot table module
}
\description{
Generate the UI for a pivot table
}
\examples{
ui <- fluidPage(
      pivot_module_UI(id = "id1", pivot_vars = my_pivot_vars)
)
}
