% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_rate_module.R
\name{pivot_rate_module}
\alias{pivot_rate_module}
\title{The server function for a pivot table module}
\usage{
pivot_rate_module(input, output, session, ns_id, numer_df, denom_df,
  pivot_vars, record_limit = 1e+06, show_filter_text = T)
}
\arguments{
\item{input}{A standard argument used by shiny when creating the module.}

\item{output}{A standard argument used by shiny when creating the module.}

\item{session}{A standard argument used by shiny when creating the module.}

\item{ns_id}{The module namespace id as a string. Must match a namespace id of the corresponding UI module element.}

\item{pivot_vars}{A table constructed using the get_pivot_vars function.}

\item{record_limit}{The maximum number of rows to bring into R to display. This is a saftely measure. You probably don't want to bring 100 million rows of data into R from a database. Defaults to 1 million.}

\item{df}{A local dataframe/tibble or tbl_dbi database connection object.}
}
\value{
The server function needed for a pivot table module.
}
\description{
This function should be passed to callModule. See example.
}
\examples{
 # note that the namespace id must
server <- function(input, output, session){
   callModule(pivot_module, id = "id1", ns_id = "id1", df = df1, pivot_vars = pivot_vars1, record_limit = 20)
}

}
