% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_module.R
\name{get_pivot_vars}
\alias{get_pivot_vars}
\title{Create a table with one row per pivot variable}
\usage{
get_pivot_vars(df, max_levels = 1000)
}
\arguments{
\item{df}{A local dataframe or tbl_dbi database table}

\item{max_levels}{The maximum number of levels a pivot variable is allowed to have.}
}
\value{
A tibble with one row per pivot variable with the variable's name, number of levels, and a list column containing the levels.
}
\description{
Create a table with one row per pivot variable
}
\examples{
# local table
df1 <- starwars \%>\%
     select_if(is.character)
pivot_vars1 <- get_pivot_vars(df1)

# using a database
con <- DBI::dbConnect(RSQLite::SQLite(), path = ":memory:")
copy_to(con, df1, "star_wars")
df2 <- tbl(con, "star_wars")
pivot_vars2 <- get_pivot_vars(df2)

}
